[CONFIGURATION] move metrics SDK component builders out of configuration_core
