+++ sdk/src/configuration/BUILD
@@ -24,7 +24,6 @@ cc_library(
         "//sdk:headers",
         "//sdk/src/common:wildcard_match",
         "//sdk/src/logs",
-        "//sdk/src/metrics",
         "//sdk/src/resource",
         "//sdk/src/trace",
     ],
@@ -56,7 +55,10 @@ cc_library(
 
 cc_library(
     name = "configuration_metrics_builders",
-    srcs = ["metrics_builders.cc"],
+    srcs = [
+        "metrics_builder_utils.cc",
+        "metrics_builders.cc",
+    ],
     include_prefix = "src/configuration",
     deps = [
         ":configuration_core",
