+++ sdk/src/configuration/CMakeLists.txt
@@ -28,7 +28,7 @@ target_include_directories(
 target_link_libraries(
   opentelemetry_configuration_core
   PUBLIC opentelemetry_api opentelemetry_common opentelemetry_resources
-         opentelemetry_trace opentelemetry_metrics opentelemetry_logs)
+         opentelemetry_trace opentelemetry_logs)
 
 # Trace SDK component builders
 add_library(opentelemetry_configuration_trace_builders trace_builders.cc)
@@ -63,7 +63,8 @@ target_link_libraries(
   PUBLIC opentelemetry_logs opentelemetry_configuration_core)
 
 # Metrics SDK component builders
-add_library(opentelemetry_configuration_metrics_builders metrics_builders.cc)
+add_library(opentelemetry_configuration_metrics_builders
+            metrics_builder_utils.cc metrics_builders.cc)
 
 set_target_properties(opentelemetry_configuration_metrics_builders
                       PROPERTIES EXPORT_NAME configuration_metrics_builders)
@@ -93,9 +94,9 @@ target_include_directories(
 target_link_libraries(
   opentelemetry_configuration_registry_factory
   PUBLIC opentelemetry_configuration_core
-  PRIVATE opentelemetry_configuration_trace_builders
-          opentelemetry_configuration_metrics_builders
-          opentelemetry_configuration_logs_builders)
+         opentelemetry_configuration_trace_builders
+         opentelemetry_configuration_metrics_builders
+         opentelemetry_configuration_logs_builders)
 
 if(OPENTELEMETRY_INSTALL)
   opentelemetry_add_pkgconfig(
